% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{config.cravat}
\alias{config.cravat}
\title{config.cravat}
\usage{
config.cravat(fmt = "yaml", to = "stdout")
}
\arguments{
\item{fmt}{Format of output. json or yaml.}

\item{to}{"stdout" to print. "return" to return}
}
\description{
config.cravat
}
