% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{config.system}
\alias{config.system}
\title{config.system}
\usage{
config.system(fmt = "json", to = "return", quiet = TRUE)
}
\value{
A named list. OakVar system config information
}
\description{
config.system
}
\examples{
# Get the named list of the OakVar system configuration
#roakvar::config.system()

}
