% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{config.user}
\alias{config.user}
\title{config.user}
\usage{
config.user(fmt = "json", to = "return", quiet = TRUE)
}
\value{
A named list. OakVar user config information
}
\description{
config.user
}
\examples{
# Get the named list of the OakVar user configuration
#roakvar::config.user()

}
