% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{gui}
\alias{gui}
\title{gui}
\usage{
gui(
  servermode = FALSE,
  headless = FALSE,
  http_only = FALSE,
  debug = FALSE,
  result = NULL,
  webapp = NULL,
  port = NULL,
  noguest = FALSE
)
}
\arguments{
\item{servermode}{Runs in multiuser mode}

\item{headless}{do not open the OakVar web page}

\item{http_only}{Force not to accept https connection}

\item{debug}{Console echoes exceptions written to log file.}

\item{result}{Path to a OakVar result SQLite file}

\item{webapp}{Name of OakVar webapp module to run}

\item{port}{Port number for OakVar graphical user interface}

\item{noguest}{Diasbles guest mode}
}
\description{
gui
}
