% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.install}
\alias{module.install}
\title{module.install}
\usage{
module.install(
  modules = NULL,
  force = FALSE,
  force_data = FALSE,
  yes = FALSE,
  skip_dependencies = FALSE,
  private = FALSE,
  skip_data = FALSE,
  md = NULL,
  to = "return",
  quiet = TRUE
)
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
Installs OakVar modules.
}
\examples{
# Install the ClinVar module without confirmation
#roakvar::module.install(modules="clinvar", force=True)

}
