% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.installbase}
\alias{module.installbase}
\title{module.installbase}
\usage{
module.installbase(force = NULL, force_data = FALSE, md = NULL, quiet = TRUE)
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
installs base modules.
}
\examples{
# Install OakVar system modules
#roakvar::module.installbase()

}
