% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.ls}
\alias{module.ls}
\title{module.ls}
\usage{
module.ls(
  pattern = ".*",
  available = FALSE,
  types = list(),
  include_hidden = FALSE,
  tags = list(),
  nameonly = FALSE,
  raw_bytes = FALSE,
  md = NULL,
  fmt = NULL,
  to = "return",
  quiet = TRUE
)
}
\value{
A named list. List of modules
}
\description{
lists modules.
}
\examples{
# Get the list of all installed modules
#roakvar::module.ls()
# Get the list of all available modules
#roakvar::module.ls(available=TRUE)
# Get the list of all available modules of the type "converter"
#roakvar::module.ls(available=TRUE, types="converter")

}
