% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.pack}
\alias{module.pack}
\title{module.pack}
\usage{
module.pack(module = NULL, outdir = ".", quiet = TRUE)
}
\value{
A boolean. A boolean. TRUE if successful, FALSE if not
}
\description{
module.pack
}
\examples{
# Pack a module "mymodule" into one zip file for its code and another zip file for its data.
#roakvar::store.pack(module="mymodule")

}
