% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.uninstall}
\alias{module.uninstall}
\title{module.uninstall}
\usage{
module.uninstall(modules = NULL, yes = FALSE, md = NULL, quiet = TRUE)
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
module.uninstall
}
\examples{
# Uninstall the ClinVar module without confirmation
#roakvar::module.uninstall(modules="clinvar", force=True)

}
