% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.update}
\alias{module.update}
\title{module.update}
\usage{
module.update(
  modules = NULL,
  y = FALSE,
  strategy = "consensus",
  md = NULL,
  quiet = TRUE
)
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
updates modules.
}
\examples{
# Update the ClinVar module without confirmation
#roakvar::module.update(modules="clinvar", force=True)

}
