% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{new.annotator}
\alias{new.annotator}
\title{new.annotator}
\usage{
new.annotator(annotator_name = "exampleannotator", md = NULL, quiet = TRUE)
}
\value{
A string. Location of the new annotator module
}
\description{
new.annotator
}
\examples{
# Create an annotator template at the OakVar modules directory/annotators/annotatortest
#roakvar::new.annotator(annotator_name="annotatortest")

}
