% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{new.exampleinput}
\alias{new.exampleinput}
\title{new.exampleinput}
\usage{
new.exampleinput(directory = ".", quiet = TRUE)
}
\value{
A string. Location of the example input file
}
\description{
new.exampleinput
}
\examples{
# Create an example input file in the current working directory
#roakvar::new.exampleinput()
# Create an example input file at /home/user1/
#roakvar::new.exampleinput(directory="/home/user1")

}
