% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.config.cravat}
\alias{ov.config.cravat}
\title{ov.config.cravat}
\usage{
ov.config.cravat(fmt = "yaml", to = "stdout")
}
\arguments{
\item{fmt}{Format of output. json or yaml.}

\item{to}{"stdout" to print. "return" to return}
}
\description{
ov.config.cravat
}
