% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.module.info}
\alias{ov.module.info}
\title{ov.module.info}
\usage{
ov.module.info(
  module = NULL,
  local = FALSE,
  md = NULL,
  fmt = "json",
  to = "return",
  quiet = TRUE
)
}
\arguments{
\item{module}{Module to get info about}

\item{local}{Include local info}

\item{md}{Specify the root directory of OakVar modules}

\item{fmt}{format of module information data. json or yaml}

\item{to}{"stdout" to stdout / "return" to return}

\item{quiet}{Run quietly}
}
\value{
A named list. Information of the queried module
}
\description{
ov.module.info
}
\examples{
# Get the information of the ClinVar module
ov.module.info(module="clinvar")

}
