% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.module.install}
\alias{ov.module.install}
\title{ov.module.install}
\usage{
ov.module.install(
  modules = NULL,
  force = FALSE,
  force_data = FALSE,
  yes = FALSE,
  skip_dependencies = FALSE,
  private = FALSE,
  skip_data = FALSE,
  md = NULL,
  to = "return",
  quiet = TRUE
)
}
\arguments{
\item{modules}{Modules to install. May be regular expressions.}

\item{force}{Install module even if latest version is already installed}

\item{force_data}{Download data even if latest data is already installed}

\item{yes}{Proceed without prompt}

\item{skip_dependencies}{Skip installing dependencies}

\item{private}{Install a private module}

\item{skip_data}{Skip installing data}

\item{md}{Specify the root directory of OakVar modules}

\item{to}{'stdout' to print. 'return' to return}

\item{quiet}{suppress stdout output}
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
Installs OakVar modules.
}
\examples{
# Install the ClinVar module
ov.module.install(modules="clinvar")
# Install the ClinVar and the COSMIC modules
ov.module.install(modules=list("clinvar", "cosmic")
# Re-install the ClinVar module overwriting the already installed copy
ov.module.install(modules="clinvar", force=TRUE)

}
