% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.module.ls}
\alias{ov.module.ls}
\title{ov.module.ls}
\usage{
ov.module.ls(
  pattern = ".*",
  available = FALSE,
  types = list(),
  include_hidden = FALSE,
  tags = list(),
  nameonly = FALSE,
  raw_bytes = FALSE,
  md = NULL,
  fmt = "json",
  to = "return",
  quiet = TRUE
)
}
\arguments{
\item{pattern}{Regular expression for module names}

\item{available}{Include available modules}

\item{types}{Only list modules of certain types}

\item{include_hidden}{Include hidden modules}

\item{tags}{Only list modules of given tag(s)}

\item{nameonly}{Only list module names}

\item{raw_bytes}{Machine readable data sizes}

\item{md}{Specify the root directory of OakVar modules}

\item{fmt}{Output format. tabular or json}

\item{to}{stdout to print / return to return}

\item{quiet}{Run quietly}
}
\value{
A named list. List of modules
}
\description{
lists modules.
}
\examples{
# Get the list of all installed modules
ov.module.ls()
# Get the list of all available modules
ov.module.ls(available=TRUE)
# Get the list of all available modules of the type "converter"
ov.module.ls(available=TRUE, types="converter")

}
