% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.module.uninstall}
\alias{ov.module.uninstall}
\title{ov.module.uninstall}
\usage{
ov.module.uninstall(modules = NULL, yes = FALSE, md = NULL, quiet = TRUE)
}
\arguments{
\item{modules}{Modules to uninstall}

\item{yes}{Proceed without prompt}

\item{md}{Specify the root directory of OakVar modules}

\item{quiet}{Run quietly}
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
ov.module.uninstall
}
\examples{
# Uninstall the ClinVar module
ov.module.uninstall(modules="clinvar")# Uninstall the ClinVar and the COSMIC modules
ov.module.uninstall(modules=("clinvar", "cosmic")

}
