% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.module.update}
\alias{ov.module.update}
\title{ov.module.update}
\usage{
ov.module.update(
  modules = NULL,
  y = FALSE,
  strategy = "consensus",
  md = NULL,
  quiet = TRUE
)
}
\arguments{
\item{modules}{Modules to update.}

\item{y}{Proceed without prompt}

\item{strategy}{Dependency resolution strategy. "consensus" will attempt to resolve dependencies. "force" will install the highest available version. "skip" will skip modules with constraints.}

\item{md}{Specify the root directory of OakVar modules}

\item{quiet}{suppress stodout output}
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
updates modules.
}
\examples{
# Update the ClinVar module
ov.module.update(modules="clinvar")
# Update all the installed modules
ov.module.update()

}
