% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.new.annotator}
\alias{ov.new.annotator}
\title{ov.new.annotator}
\usage{
ov.new.annotator(annotator_name = "annotator", md = NULL)
}
\arguments{
\item{annotator_name}{Annotator name}

\item{md}{Specify the root directory of OakVar modules}
}
\description{
ov.new.annotator
}
