% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.new.exampleinput}
\alias{ov.new.exampleinput}
\title{ov.new.exampleinput}
\usage{
ov.new.exampleinput(directory = ".")
}
\arguments{
\item{directory}{Directory to make the example input file in}
}
\description{
ov.new.exampleinput
}
