% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.report}
\alias{ov.report}
\title{ov.report}
\usage{
ov.report(
  dbpath = NULL,
  reporttypes = list(),
  filterpath = NULL,
  filter = NULL,
  filtersql = NULL,
  filtername = NULL,
  filterstring = NULL,
  savepath = NULL,
  confpath = NULL,
  module_name = NULL,
  nogenelevelonvariantlevel = FALSE,
  confs = "{}",
  inputfiles = NULL,
  separatesample = FALSE,
  output_dir = NULL,
  do_not_change_status = FALSE,
  silent = FALSE,
  system_option = NULL,
  module_option = NULL,
  concise_report = FALSE,
  includesample = NULL,
  excludesample = NULL,
  package = NULL,
  md = NULL
)
}
\arguments{
\item{dbpath}{Path to aggregator output}

\item{reporttypes}{report types}

\item{filterpath}{Path to filter file}

\item{filter}{==SUPPRESS==}

\item{filtersql}{Filter SQL}

\item{filtername}{Name of filter (stored in aggregator output)}

\item{filterstring}{==SUPPRESS==}

\item{savepath}{Path to save file}

\item{confpath}{path to a conf file}

\item{module_name}{report module name}

\item{nogenelevelonvariantlevel}{Use this option to prevent gene level result from being added to variant level result.}

\item{confs}{Configuration string}

\item{inputfiles}{Original input file path}

\item{separatesample}{Write each variant-sample pair on a separate line}

\item{output_dir}{directory for output files}

\item{do_not_change_status}{Job status in status.json will not be changed}

\item{silent}{Suppress output to STDOUT}

\item{system_option}{System option in key=value syntax. For example, --system-option modules_dir=/home/user/oakvar/modules}

\item{module_option}{Module-specific option in module_name.key=value syntax. For example, --module-option vcfreporter.type=separate}

\item{concise_report}{Generate concise report with default columns defined by annotation modules}

\item{includesample}{Sample IDs to include}

\item{excludesample}{Sample IDs to exclude}

\item{package}{Use filters and report types in a package}

\item{md}{Specify the root directory of OakVar modules (annotators, etc)}
}
\description{
ov.report
}
