% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.run}
\alias{ov.run}
\title{ov.run}
\usage{
ov.run(
  inputs = NULL,
  annotators = list(),
  annotators_replace = list(),
  excludes = list(),
  run_name = NULL,
  output_dir = NULL,
  startat = NULL,
  endat = NULL,
  skip = NULL,
  conf = "oc.yml",
  confs = NULL,
  verbose = NULL,
  reports = list(),
  genome = NULL,
  cleandb = FALSE,
  newlog = NULL,
  note = NULL,
  mp = NULL,
  forcedinputformat = NULL,
  temp_files = NULL,
  writeadmindb = NULL,
  jobid = NULL,
  show_version = NULL,
  separatesample = NULL,
  unique_variants = NULL,
  primary_transcript = list(),
  clean_run = NULL,
  do_not_change_status = NULL,
  module_option = NULL,
  system_option = NULL,
  silent = NULL,
  concise_report = NULL,
  package = NULL,
  filtersql = NULL,
  includesample = NULL,
  excludesample = NULL,
  filter = NULL,
  filterpath = NULL,
  md = NULL,
  mapper_name = list(),
  postaggregators = list()
)
}
\arguments{
\item{inputs}{Input file(s). One or more variant files in a supported format like VCF.  See the -i/--input-format flag for supported formats. In the special case where you want to add annotations to an existing OakVar analysis, provide the output sqlite database from the previous run as input instead of a variant input file.}

\item{annotators}{Annotator module names or directories. If --package is used also, annotator modules defined with -a will be added.}

\item{annotators_replace}{Annotator module names or directories. If --package is used also, annotator modules defined with -A will replace those defined with --package. -A has priority over -a.}

\item{excludes}{annotators to exclude}

\item{run_name}{name of oakvar run}

\item{output_dir}{directory for output files}

\item{startat}{starts at given stage}

\item{endat}{ends after given stage.}

\item{skip}{skips given stage(s).}

\item{conf}{path to a conf file}

\item{confs}{configuration string}

\item{verbose}{verbose}

\item{reports}{Reporter types or reporter module directories}

\item{genome}{reference genome of input. OakVar will lift over to hg38 if needed.}

\item{cleandb}{deletes the existing result database and creates a new one.}

\item{newlog}{deletes the existing log file and creates a new one.}

\item{note}{note will be written to the run status file (.status.json)}

\item{mp}{number of processes to use to run annotators}

\item{forcedinputformat}{Force input format}

\item{temp_files}{Leave temporary files after run is complete.}

\item{writeadmindb}{Write job information to admin db after job completion}

\item{jobid}{Job ID for server version}

\item{show_version}{Shows OakVar version.}

\item{separatesample}{Separate variant results by sample}

\item{unique_variants}{Set to get only unique variants in output}

\item{primary_transcript}{"mane" for MANE transcripts as primary transcripts, or a path to a file of primary transcripts. MANE is default.}

\item{clean_run}{Deletes all previous output files for the job and generate new ones.}

\item{do_not_change_status}{Job status in status.json will not be changed}

\item{module_option}{Module-specific option in module_name.key=value syntax. For example, --module-option vcfreporter.type=separate}

\item{system_option}{System option in key=value syntax. For example, --system-option modules_dir=/home/user/oakvar/modules}

\item{silent}{Runs silently.}

\item{concise_report}{Generate concise reports with default columns defined by each annotation module}

\item{package}{Use package}

\item{filtersql}{Filter SQL}

\item{includesample}{Sample IDs to include}

\item{excludesample}{Sample IDs to exclude}

\item{filter}{==SUPPRESS==}

\item{filterpath}{Path to a filter file}

\item{md}{Specify the root directory of OakVar modules (annotators, etc)}

\item{mapper_name}{Mapper module name or mapper module directory}

\item{postaggregators}{Postaggregators to run. Additionally, tagsampler, casecontrol, varmeta, and vcfinfo will automatically run depending on conditions.}
}
\description{
Run a job
}
