% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.store.checklogin}
\alias{ov.store.checklogin}
\title{ov.store.checklogin}
\usage{
ov.store.checklogin(username = NULL, password = NULL)
}
\arguments{
\item{username}{username}

\item{password}{password}
}
\description{
ov.store.checklogin
}
