% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.store.createaccount}
\alias{ov.store.createaccount}
\title{ov.store.createaccount}
\usage{
ov.store.createaccount(username = NULL, password = NULL, quiet = TRUE)
}
\arguments{
\item{username}{use your email as your username.}

\item{password}{this is your password.}

\item{quiet}{Run quietly}
}
\value{
A string. Response from the store server
}
\description{
ov.store.createaccount
}
\examples{
# Create a store account
ov.store.newaccount(username="user1", password="password")

}
