% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.store.newaccount}
\alias{ov.store.newaccount}
\title{ov.store.newaccount}
\usage{
ov.store.newaccount(username = NULL, password = NULL)
}
\arguments{
\item{username}{use your email as your username.}

\item{password}{this is your password.}
}
\description{
ov.store.newaccount
}
