% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.store.publish}
\alias{ov.store.publish}
\title{ov.store.publish}
\usage{
ov.store.publish(
  module = NULL,
  data = FALSE,
  code = FALSE,
  user = NULL,
  password = NULL,
  force_yes = FALSE,
  overwrite = FALSE,
  md = NULL
)
}
\arguments{
\item{module}{module to publish}

\item{data}{publishes module with data.}

\item{code}{publishes module without data.}

\item{user}{user to publish as. Typically your email.}

\item{password}{password for the user. Enter at prompt if missing.}

\item{force_yes}{overrides yes to overwrite question}

\item{overwrite}{overwrites a published module/version}

\item{md}{Specify the root directory of OakVar modules}
}
\description{
ov.store.publish
}
