% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.store.resetpassword}
\alias{ov.store.resetpassword}
\title{ov.store.resetpassword}
\usage{
ov.store.resetpassword(username = NULL)
}
\arguments{
\item{username}{username}
}
\description{
ov.store.resetpassword
}
