% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.store.verifyemail}
\alias{ov.store.verifyemail}
\title{ov.store.verifyemail}
\usage{
ov.store.verifyemail(username = NULL, quiet = TRUE)
}
\arguments{
\item{username}{username}

\item{quiet}{Run quietly}
}
\value{
\code{NULL}
}
\description{
ov.store.verifyemail
}
\examples{
# Ask the store to send an email to verify the email of a user account
ov.store.verifyemail(username="user1")

}
