% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.system.config}
\alias{ov.system.config}
\title{ov.system.config}
\usage{
ov.system.config(fmt = "json", to = "return", quiet = TRUE)
}
\arguments{
\item{fmt}{Format of output. json or yaml.}

\item{to}{"stdout" to print. "return" to return}

\item{quiet}{Run quietly}
}
\value{
A named list. System config information
}
\description{
ov.system.config
}
\examples{
# Get named list of the OakVar system configuration
ov.system.config()
# Get the OakVar system configuration in YAML text
ov.system.config(fmt="yaml")# Print to stdout the OakVar system configuration in YAML text
ov.system.config(fmt="yaml", to="stdout")

}
