% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.system.md}
\alias{ov.system.md}
\title{ov.system.md}
\usage{
ov.system.md(directory = NULL, to = "return", quiet = TRUE)
}
\arguments{
\item{directory}{sets modules directory.}

\item{to}{'stdout' to print. 'return' to return.}

\item{quiet}{Run quietly}
}
\value{
A string. OakVar modules directory
}
\description{
displays or changes OakVar modules directory.
}
\examples{
# Get the OakVar modules directory
ov.system.md()
# Set the OakVar modules directory to /home/user1/.oakvar/modules
ov.system.md(directory="/home/user1/.oakvar/modules")

}
