% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.system.setup}
\alias{ov.system.setup}
\title{ov.system.setup}
\usage{
ov.system.setup(setup_file = NULL, quiet = TRUE)
}
\arguments{
\item{setup_file}{setup file to use}

\item{quiet}{Run quietly}
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
Sets up OakVar system
}
\examples{
# Set up OakVar with defaults
ov.system.setup()
# Set up OakVar with a setup file
ov.system.setup(setup_file="setup.yml")

}
