% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.util.addjob}
\alias{ov.util.addjob}
\title{ov.util.addjob}
\usage{
ov.util.addjob(path = NULL, user = "default")
}
\arguments{
\item{path}{Path to result database}

\item{user}{User who will own the job. Defaults to single user default user.}
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
ov.util.addjob
}
\examples{
# Add a result file to the job list of a user
ov.util.addjob(path="example.sqlite", user="user1")

}
