% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.util.mergesqlite}
\alias{ov.util.mergesqlite}
\title{ov.util.mergesqlite}
\usage{
ov.util.mergesqlite(path = NULL, outpath = NULL)
}
\arguments{
\item{path}{Path to result database}

\item{outpath}{Output SQLite file path}
}
\description{
ov.util.mergesqlite
}
