% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.util.sendgui}
\alias{ov.util.sendgui}
\title{ov.util.sendgui}
\usage{
ov.util.sendgui(path = NULL, user = "default")
}
\arguments{
\item{path}{Path to result database}

\item{user}{User who will own the job. Defaults to single user default user.}
}
\description{
ov.util.sendgui
}
