% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.util.sqliteinfo}
\alias{ov.util.sqliteinfo}
\title{ov.util.sqliteinfo}
\usage{
ov.util.sqliteinfo(paths = NULL, fmt = "json", to = "return")
}
\arguments{
\item{paths}{SQLite result file paths}

\item{fmt}{Output format. text / json / yaml}

\item{to}{Output to. stdout / return}
}
\value{
A named list. Information of a job SQLite file
}
\description{
ov.util.sqliteinfo
}
\examples{
# Get the named list of the information of an analysis result file
ov.util.sqliteinfo(paths="example.sqlite")

}
