% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.util.updateresult}
\alias{ov.util.updateresult}
\title{ov.util.updateresult}
\usage{
ov.util.updateresult(dbpath = NULL, recursive = FALSE, backup = FALSE)
}
\arguments{
\item{dbpath}{path to a result db file or a directory}

\item{recursive}{recursive operation}

\item{backup}{backup original copy with .bak extension}
}
\description{
ov.util.updateresult
}
