% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{report}
\alias{report}
\title{report}
\usage{
report(
  dbpath = NULL,
  reports = list(),
  filterpath = NULL,
  filter = NULL,
  filtersql = NULL,
  filtername = NULL,
  filterstring = NULL,
  savepath = NULL,
  confpath = NULL,
  module_name = NULL,
  nogenelevelonvariantlevel = FALSE,
  confs = "{}",
  inputfiles = NULL,
  separatesample = FALSE,
  output_dir = NULL,
  do_not_change_status = FALSE,
  quiet = TRUE,
  system_option = NULL,
  module_option = NULL,
  concise_report = FALSE,
  includesample = NULL,
  excludesample = NULL,
  package = NULL,
  md = NULL,
  level = NULL
)
}
\value{
A string, a named list, or a dataframe. Output of reporters
}
\description{
Generate reports from a job
}
\examples{
# Generate a CSV-format report file from the job result file example.sqlite
#roakvar::report(dbpath="example.sqlite", reports="csv")

}
