% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.account.change}
\alias{store.account.change}
\title{store.account.change}
\usage{
store.account.change(newpw = NULL, quiet = TRUE)
}
\value{
A string. Response from the store server
}
\description{
store.account.change
}
\examples{
# Change the password of a store account
#roakvar::store.changepassword(email="user1", 
# current_password="password", new_password="newpassword")

}
