% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.account.check}
\alias{store.account.check}
\title{store.account.check}
\usage{
store.account.check(quiet = TRUE)
}
\value{
A boolean. A boolean. TRUE if successful, FALSE if not
}
\description{
store.account.check
}
\examples{
# Check if the current is logged in the OakVar Store nor not. 
#roakvar::store.checklogin(email="user1", password="password")

}
