% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.account.create}
\alias{store.account.create}
\title{store.account.create}
\usage{
store.account.create(email = NULL, pw = NULL, quiet = TRUE)
}
\value{
A boolean. TRUE if successful, FALSE if not.
}
\description{
store.account.create
}
\examples{
# Create a store account
#roakvar::store.createaccount(email="user1", password="password")

}
