% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.account.delete}
\alias{store.account.delete}
\title{store.account.delete}
\usage{
store.account.delete(quiet = TRUE)
}
\value{
A boolean. TRUE if successful, FALSE if not.
}
\description{
store.account.delete
}
\examples{
# Create a store account
#roakvar::store.deleteaccount(email="user1", password="password")

}
