% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.account.login}
\alias{store.account.login}
\title{store.account.login}
\usage{
store.account.login(email = NULL, pw = NULL, quiet = TRUE)
}
\value{
\code{NULL}
}
\description{
store.account.login
}
\examples{
# Log in to the OakVar Store
#roakvar::store.account.login(email="user1", pw="password")

}
