% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.account.logout}
\alias{store.account.logout}
\title{store.account.logout}
\usage{
store.account.logout(quiet = TRUE)
}
\value{
\code{NULL}
}
\description{
store.account.logout
}
\examples{
# Log out from the OakVar Store
#roakvar::store.account.logout()

}
