% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.account.reset}
\alias{store.account.reset}
\title{store.account.reset}
\usage{
store.account.reset(quiet = TRUE, email = NULL)
}
\value{
A boolean. A boolean. TRUE if successful, FALSE if not
}
\description{
store.account.reset
}
\examples{
# Ask the store to send an email to reset the password of a store account
#roakvar::store.account.reset(email="user1")

}
