% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.changepassword}
\alias{store.changepassword}
\title{store.changepassword}
\usage{
store.changepassword(
  username = NULL,
  current_password = NULL,
  new_password = NULL
)
}
\arguments{
\item{username}{username}

\item{current_password}{current password}

\item{new_password}{new password}
}
\description{
store.changepassword
}
