% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.checklogin}
\alias{store.checklogin}
\title{store.checklogin}
\usage{
store.checklogin(username = NULL, password = NULL)
}
\arguments{
\item{username}{username}

\item{password}{password}
}
\description{
store.checklogin
}
