% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.fetch}
\alias{store.fetch}
\title{store.fetch}
\usage{
store.fetch(quiet = TRUE, email = NULL, pw = NULL, clean = FALSE)
}
\value{
A boolean. A boolean. TRUE if successful, FALSE if not
}
\description{
store.fetch
}
\examples{
# Fetch the store information
#roakvar::store.fetch()

}
