% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.newaccount}
\alias{store.newaccount}
\title{store.newaccount}
\usage{
store.newaccount(username = NULL, password = NULL)
}
\arguments{
\item{username}{use your email as your username.}

\item{password}{this is your password.}
}
\description{
store.newaccount
}
