% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.oc.changepw}
\alias{store.oc.changepw}
\title{store.oc.changepw}
\usage{
store.oc.changepw(username = NULL, cur_pw = NULL, new_pw = NULL)
}
\value{
A boolean. TRUE if successful, FALSE if not.
}
\description{
store.oc.changepw
}
\examples{
# Change the password of a developer account at the OpenCRAVAT store
#roakvar::store.oc.changepw(username="user1", cur_pw="password", new_pw="newpassword")

}
