% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.oc.checklogin}
\alias{store.oc.checklogin}
\title{store.oc.checklogin}
\usage{
store.oc.checklogin(username = NULL, password = NULL)
}
\value{
A boolean. TRUE if successful, FALSE if not.
}
\description{
store.oc.checklogin
}
\examples{
# Check the username and password of a developer account at the OpenCRAVAT store
#roakvar::store.oc.checklogin(username="user1")

}
