% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.oc.newaccount}
\alias{store.oc.newaccount}
\title{store.oc.newaccount}
\usage{
store.oc.newaccount(username = NULL, password = NULL)
}
\value{
A boolean. TRUE if successful, FALSE if not.
}
\description{
store.oc.newaccount
}
\examples{
# Create a developer account at the OpenCRAVAT store
#roakvar::store.oc.newaccount(username="user1", password="password")

}
