% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.oc.publish}
\alias{store.oc.publish}
\title{store.oc.publish}
\usage{
store.oc.publish(
  module = NULL,
  data = FALSE,
  code = FALSE,
  user = NULL,
  password = NULL,
  force_yes = FALSE,
  overwrite = FALSE,
  md = NULL,
  quiet = TRUE
)
}
\value{
A boolean. TRUE if successful, FALSE if not.
}
\description{
store.oc.publish
}
\examples{
# Publish a module to the OpenCRAVAT store
#roakvar::store.oc.publish(module="clinvar", user="user1", password="password", code=TRUE)

}
