% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.oc.resetpw}
\alias{store.oc.resetpw}
\title{store.oc.resetpw}
\usage{
store.oc.resetpw(username = NULL)
}
\value{
A boolean. TRUE if successful, FALSE if not.
}
\description{
store.oc.resetpw
}
\examples{
# Reset the password of a developer account at the OpenCRAVAT store
#roakvar::store.oc.resetpw(username="user1")

}
