% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.oc.verifyemail}
\alias{store.oc.verifyemail}
\title{store.oc.verifyemail}
\usage{
store.oc.verifyemail(username = NULL)
}
\value{
A boolean. TRUE if successful, FALSE if not.
}
\description{
store.oc.verifyemail
}
\examples{
# Sends a verification email for a developer account at the OpenCRAVAT store
#roakvar::store.oc.verifyemail(username="user1")

}
