% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.register}
\alias{store.register}
\title{store.register}
\usage{
store.register(
  module_name = NULL,
  md = NULL,
  quiet = TRUE,
  code_url = NULL,
  data_url = NULL,
  overwrite = FALSE
)
}
\value{
A boolean. A boolean. TRUE if successful, FALSE if not
}
\description{
store.register
}
\examples{
# Publish "customannot" module to the store
#roakvar::store.publish(module="customannot", 
# code_url="https://test.com/customannot__1.0.0__code.zip", 
# data_url="https://test.com/customannot__1.0.0__data.zip")

}
