% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{store.url}
\alias{store.url}
\title{store.url}
\usage{
store.url(quiet = TRUE)
}
\value{
character
}
\description{
store.url
}
\examples{
# Returns the URL of the OakVar store.
#roakvar::store.account.url()

}
