% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{system.check}
\alias{system.check}
\title{system.check}
\usage{
system.check(to = "return", quiet = TRUE)
}
\value{
A boolean. true if no problem or false if not.
}
\description{
system.check
}
\examples{
# Check if OakVar is set up correctly.
#roakvar::system.check()

}
