% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{system.config}
\alias{system.config}
\title{system.config}
\usage{
system.config(fmt = "json", to = "return", quiet = TRUE)
}
\value{
A named list. System config information
}
\description{
system.config
}
\examples{
# Get named list of the OakVar system configuration
#roakvar::system.config()
# Get the OakVar system configuration in YAML text
#roakvar::system.config(fmt="yaml")# Print to stdout the OakVar system configuration in YAML text
#roakvar::system.config(fmt="yaml", to="stdout")

}
