% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{system.md}
\alias{system.md}
\title{system.md}
\usage{
system.md(directory = NULL, to = "return", quiet = TRUE)
}
\value{
A string. OakVar modules directory
}
\description{
displays or changes OakVar modules directory.
}
\examples{
# Get the OakVar modules directory
#roakvar::system.md()
# Set the OakVar modules directory to /home/user1/.oakvar/modules
#roakvar::system.md(directory="/home/user1/.oakvar/modules")

}
