% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{system.setup}
\alias{system.setup}
\title{system.setup}
\usage{
system.setup(setup_file = NULL, clean = FALSE, quiet = TRUE)
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
Sets up OakVar system
}
\examples{
# Set up OakVar with defaults
#roakvar::system.setup()
# Set up OakVar with a setup file
#roakvar::system.setup(setup_file="setup.yml")

}
