% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{util.addjob}
\alias{util.addjob}
\title{util.addjob}
\usage{
util.addjob(path = NULL, user = "default")
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
util.addjob
}
\examples{
# Add a result file to the job list of a user
#roakvar::util.addjob(path="example.sqlite", user="user1")

}
