% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{util.filtersqlite}
\alias{util.filtersqlite}
\title{util.filtersqlite}
\usage{
util.filtersqlite(
  paths = NULL,
  out = ".",
  suffix = "filtered",
  filterpath = NULL,
  filtersql = NULL,
  includesample = NULL,
  excludesample = NULL
)
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
util.filtersqlite
}
\examples{
# Filter an analysis result file with an SQL filter set
#roakvar::util.filtersqlite(paths="example.sqlite", 
#  filtersql='base__so=="MIS" and gnomad__af>0.01')
# Filter two analysis result files with a filter definition file
#roakvar::util.filtersqlite(paths=list("example1.sqlite", 
#  "example2.sqlite"), filterpath="filter.json")

}
