% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{util.mergesqlite}
\alias{util.mergesqlite}
\title{util.mergesqlite}
\usage{
util.mergesqlite(path = NULL, outpath = NULL)
}
\arguments{
\item{path}{Path to result database}

\item{outpath}{Output SQLite file path}
}
\description{
util.mergesqlite
}
