% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{util.sendgui}
\alias{util.sendgui}
\title{util.sendgui}
\usage{
util.sendgui(path = NULL, user = "default")
}
\arguments{
\item{path}{Path to result database}

\item{user}{User who will own the job. Defaults to single user default user.}
}
\description{
util.sendgui
}
