% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{util.showsqliteinfo}
\alias{util.showsqliteinfo}
\title{util.showsqliteinfo}
\usage{
util.showsqliteinfo(paths = NULL, fmt = "text", to = "stdout")
}
\arguments{
\item{paths}{SQLite result file paths}

\item{fmt}{Output format. text / json / yaml}

\item{to}{Output to. stdout / return}
}
\description{
util.showsqliteinfo
}
