% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{util.sqliteinfo}
\alias{util.sqliteinfo}
\title{util.sqliteinfo}
\usage{
util.sqliteinfo(paths = NULL, fmt = "json", to = "return")
}
\value{
A named list. Information of a job SQLite file
}
\description{
util.sqliteinfo
}
\examples{
# Get the named list of the information of an analysis result file
#roakvar::util.sqliteinfo(paths="example.sqlite")

}
