% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{util.test}
\alias{util.test}
\title{util.test}
\usage{
util.test(rundir = NULL, modules = NULL, mod_types = NULL)
}
\arguments{
\item{rundir}{Directory for output}

\item{modules}{Name of module(s) to test. (e.g. gnomad)}

\item{mod_types}{Type of module(s) to test (e.g. annotators)}
}
\description{
util.test
}
