% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{version}
\alias{version}
\title{version}
\usage{
version(to = "return", quiet = TRUE)
}
\value{
A string. OakVar version
}
\description{
version
}
\examples{
# Get the version of the installed OakVar
#roakvar::version()

}
