% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{config.system}
\alias{config.system}
\title{config.system}
\usage{
config.system(fmt = "yaml", to = "stdout")
}
\arguments{
\item{fmt}{Format of output. json or yaml.}

\item{to}{"stdout" to print. "return" to return}
}
\description{
config.system
}
