% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{gui}
\alias{gui}
\title{gui}
\usage{
gui(
  servermode = FALSE,
  headless = FALSE,
  http_only = FALSE,
  debug = FALSE,
  result = NULL,
  webapp = NULL,
  port = NULL,
  noguest = FALSE,
  quiet = TRUE,
  to = "return"
)
}
\value{
\code{NULL}
}
\description{
gui
}
\examples{
# Launch OakVar GUI
#roakvar::gui()
# Launch OakVar Interactive Result Viewer for the OakVar analysis file example.sqlite
#roakvar::gui(result="example.sqlite")

}
