% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.info}
\alias{module.info}
\title{module.info}
\usage{
module.info(
  module = NULL,
  local = FALSE,
  md = NULL,
  fmt = "json",
  to = "return",
  quiet = TRUE
)
}
\value{
A named list. Information of the queried module
}
\description{
module.info
}
\examples{
# Get the information of the ClinVar module
#roakvar::module.info(module="clinvar")

}
