% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.install}
\alias{module.install}
\title{module.install}
\usage{
module.install(
  modules = NULL,
  force = FALSE,
  force_data = FALSE,
  yes = FALSE,
  skip_dependencies = FALSE,
  private = FALSE,
  skip_data = FALSE,
  md = NULL
)
}
\arguments{
\item{modules}{Modules to install. May be regular expressions.}

\item{force}{Install module even if latest version is already installed}

\item{force_data}{Download data even if latest data is already installed}

\item{yes}{Proceed without prompt}

\item{skip_dependencies}{Skip installing dependencies}

\item{private}{Install a private module}

\item{skip_data}{Skip installing data}

\item{md}{Specify the root directory of OakVar modules}
}
\description{
Install modules
}
