% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.installbase}
\alias{module.installbase}
\title{module.installbase}
\usage{
module.installbase(
  force = FALSE,
  force_data = FALSE,
  install_pypi_dependency = TRUE,
  md = NULL
)
}
\arguments{
\item{force}{Overwrite existing modules}

\item{force_data}{Download data even if latest data is already installed}

\item{install_pypi_dependency}{Try to install non-OakVar package dependency with pip}

\item{md}{Specify the root directory of OakVar modules}
}
\description{
module.installbase
}
