% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.ls}
\alias{module.ls}
\title{module.ls}
\usage{
module.ls(
  pattern = ".*",
  available = FALSE,
  types = list(),
  include_hidden = FALSE,
  tags = list(),
  quiet = FALSE,
  raw_bytes = FALSE,
  md = NULL,
  fmt = "tabular",
  stdout = TRUE
)
}
\arguments{
\item{pattern}{Regular expression for module names}

\item{available}{Include available modules}

\item{types}{Only list modules of certain types}

\item{include_hidden}{Include hidden modules}

\item{tags}{Only list modules of given tag(s)}

\item{quiet}{Only list module names}

\item{raw_bytes}{Machine readable data sizes}

\item{md}{Specify the root directory of OakVar modules}

\item{fmt}{Output format. tabular or json}

\item{stdout}{Print to stdout}
}
\description{
module.ls
}
