% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{module.update}
\alias{module.update}
\title{module.update}
\usage{
module.update(
  modules = NULL,
  y = FALSE,
  strategy = "consensus",
  install_pypi_dependency = TRUE,
  md = NULL
)
}
\arguments{
\item{modules}{Modules to update.}

\item{y}{Proceed without prompt}

\item{strategy}{Dependency resolution strategy. "consensus" will attemp to resolve dependencies. "force" will install the highest available version. "skip" will skip modules with constraints.}

\item{install_pypi_dependency}{Try to install non-OakVar package dependency with pip}

\item{md}{Specify the root directory of OakVar modules}
}
\description{
module.update
}
