% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{new.annotator}
\alias{new.annotator}
\title{new.annotator}
\usage{
new.annotator(annotator_name = "annotator", md = NULL)
}
\arguments{
\item{annotator_name}{Annotator name}

\item{md}{Specify the root directory of OakVar modules}
}
\description{
new.annotator
}
