% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{new.exampleinput}
\alias{new.exampleinput}
\title{new.exampleinput}
\usage{
new.exampleinput(directory = ".")
}
\arguments{
\item{directory}{Directory to make the example input file in}
}
\description{
new.exampleinput
}
