% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.module.info}
\alias{ov.module.info}
\title{ov.module.info}
\usage{
ov.module.info(module = NULL, local = FALSE, md = NULL, to = "stdout")
}
\arguments{
\item{module}{Module to get info about}

\item{local}{Include local info}

\item{md}{Specify the root directory of OakVar modules}

\item{to}{"print" to stdout / "return" to return}
}
\description{
ov.module.info
}
