% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.module.installbase}
\alias{ov.module.installbase}
\title{ov.module.installbase}
\usage{
ov.module.installbase(
  force = FALSE,
  force_data = FALSE,
  md = NULL,
  quiet = TRUE
)
}
\arguments{
\item{force}{Overwrite existing modules}

\item{force_data}{Download data even if latest data is already installed}

\item{md}{Specify the root directory of OakVar modules}

\item{quiet}{suppress stdout output}
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
installs base modules.
}
\examples{
# Install OakVar system modules
ov.module.installbase()

}
