% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.module.uninstall}
\alias{ov.module.uninstall}
\title{ov.module.uninstall}
\usage{
ov.module.uninstall(modules = NULL, yes = FALSE, md = NULL)
}
\arguments{
\item{modules}{Modules to uninstall}

\item{yes}{Proceed without prompt}

\item{md}{Specify the root directory of OakVar modules}
}
\description{
ov.module.uninstall
}
