% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.new.exampleinput}
\alias{ov.new.exampleinput}
\title{ov.new.exampleinput}
\usage{
ov.new.exampleinput(directory = ".", quiet = TRUE)
}
\arguments{
\item{directory}{Directory to make the example input file in}

\item{quiet}{Run quietly}
}
\value{
A string. Location of the example input file
}
\description{
ov.new.exampleinput
}
\examples{
# Create an example input file in the current working directory
ov.new.exampleinput()
# Create an example input file at /home/user1/
ov.new.exampleinput(directory="/home/user1")

}
