% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.store.changepassword}
\alias{ov.store.changepassword}
\title{ov.store.changepassword}
\usage{
ov.store.changepassword(
  username = NULL,
  current_password = NULL,
  new_password = NULL,
  quiet = TRUE
)
}
\arguments{
\item{username}{username}

\item{current_password}{current password}

\item{new_password}{new password}

\item{quiet}{Run quietly}
}
\value{
A string. Response from the store server
}
\description{
ov.store.changepassword
}
\examples{
# Change the password of a store account
ov.store.changepassword(username="user1", current_password="password", new_password="newpassword")

}
