% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.util.filtersqlite}
\alias{ov.util.filtersqlite}
\title{ov.util.filtersqlite}
\usage{
ov.util.filtersqlite(
  paths = NULL,
  out = ".",
  suffix = "filtered",
  filterpath = NULL,
  filtersql = NULL,
  includesample = NULL,
  excludesample = NULL
)
}
\arguments{
\item{paths}{Path to result database}

\item{out}{Output SQLite file folder}

\item{suffix}{Suffix for output SQLite files}

\item{filterpath}{Path to a filter JSON file}

\item{filtersql}{Filter SQL}

\item{includesample}{Sample IDs to include}

\item{excludesample}{Sample IDs to exclude}
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
ov.util.filtersqlite
}
\examples{
# Filter an analysis result file with an SQL filter set
ov.util.filtersqlite(paths="example.sqlite", filtersql='base__so=="MIS" and gnomad__af>0.01')# Filter two analysis result files with a filter definition file
ov.util.filtersqlite(paths=list("example1.sqlite", "example2.sqlite"), filterpath="filter.json")

}
