% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{ov.util.mergesqlite}
\alias{ov.util.mergesqlite}
\title{ov.util.mergesqlite}
\usage{
ov.util.mergesqlite(path = NULL, outpath = NULL)
}
\arguments{
\item{path}{Path to result database}

\item{outpath}{Output SQLite file path}
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
ov.util.mergesqlite
}
\examples{
# Merge two OakVar analysis result files into one SQLite file
ov.util.mergesqlite(path=list("example1.sqlite", "example2.sqlite"), outpath="merged.sqlite")

}
