% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{util.filtersqlite}
\alias{util.filtersqlite}
\title{util.filtersqlite}
\usage{
util.filtersqlite(
  paths = NULL,
  out = ".",
  suffix = "filtered",
  filterpath = NULL,
  filtersql = NULL,
  includesample = NULL,
  excludesample = NULL
)
}
\arguments{
\item{paths}{Path to result database}

\item{out}{Output SQLite file folder}

\item{suffix}{Suffix for output SQLite files}

\item{filterpath}{Path to a filter JSON file}

\item{filtersql}{Filter SQL}

\item{includesample}{Sample IDs to include}

\item{excludesample}{Sample IDs to exclude}
}
\description{
util.filtersqlite
}
