% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{util.mergesqlite}
\alias{util.mergesqlite}
\title{util.mergesqlite}
\usage{
util.mergesqlite(path = NULL, outpath = NULL)
}
\value{
A boolean. TRUE if successful, FALSE if not
}
\description{
util.mergesqlite
}
\examples{
# Merge two OakVar analysis result files into one SQLite file
#roakvar::util.mergesqlite(path=list("example1.sqlite", "example2.sqlite"), outpath="merged.sqlite")

}
