% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roakvar.R
\name{util.test}
\alias{util.test}
\title{util.test}
\usage{
util.test(
  rundir = NULL,
  modules = NULL,
  mod_types = NULL,
  to = "return",
  quiet = TRUE
)
}
\value{
A named list. Field result is a named list showing the test result for each module. Fields num_passed and num_failed show the number of passed and failed modules.
}
\description{
Test modules
}
\examples{
# Test the ClinVar module
#roakvar::util.test(modules="clinvar")
# Test the ClinVar and the COSMIC modules
#roakvar::util.test(modules=list("clinvar", "cosmic"))

}
